+++ doc/whatsnew.rst
@@ -24,7 +24,9 @@ v0.2.8.dev
 
 - Refactor tests + fix refit of :class:`pyriemann.tangentspace.TangentSpace`
 
-- Add :class:`pyriemann.clustering.PotatoField`, and an example on artifact detection
+- Add sampling SPD matrices from a Riemannian Gaussian distribution in :func:`pyriemann.datasets.sample_gaussian_spd`
+
+- Add new function :func:`pyriemann.datasets.make_gaussian_blobs` for generating random datasets with SPD matrices
 
 v0.2.7 (June 2021)
 ------------------